+++ Source/Modules/ISHServer/Install-ISHVisualBasicRuntime.ps1
@@ -33,7 +33,7 @@ function Install-ISHVisualBasicRuntime
         $osInfo=Get-ISHOSInfo
         if($osInfo.IsCore)
         {
-            # Workaround for Windows Server 2016 core
+            # Workaround for Windows Server 2016/2019 core
             # https://social.technet.microsoft.com/Forums/windowsserver/en-US/9b0f8911-07f4-420f-9e48-d31915f91528/msvbvm60dll-missing-in-core?forum=winservercore
             Write-Warning "This is a workaround for making the Visual Basic runtime available on $($osInfo.Caption) Core"